Moved more wallet method definitions
