+++ libraries/wallet/wallet.cpp
@@ -1269,124 +1269,6 @@ vector< signed_transaction > wallet_api::import_balance(
    return my->import_balance( name_or_id, wif_keys, broadcast );
 }
 
-namespace detail {
-
-vector< signed_transaction > wallet_api_impl::import_balance(
-      string name_or_id,
-      const vector<string>& wif_keys,
-      bool broadcast )
-{ try {
-   FC_ASSERT(!is_locked());
-   const dynamic_global_property_object& dpo = _remote_db->get_dynamic_global_properties();
-   account_object claimer = get_account( name_or_id );
-   uint32_t max_ops_per_tx = 30;
-
-   map< address, private_key_type > keys;  // local index of address -> private key
-   vector< address > addrs;
-   bool has_wildcard = false;
-   addrs.reserve( wif_keys.size() );
-   for( const string& wif_key : wif_keys )
-   {
-      if( wif_key == "*" )
-      {
-         if( has_wildcard )
-            continue;
-         for( const public_key_type& pub : _wallet.extra_keys[ claimer.id ] )
-         {
-            addrs.push_back( pub );
-            auto it = _keys.find( pub );
-            if( it != _keys.end() )
-            {
-               fc::optional< fc::ecc::private_key > privkey = wif_to_key( it->second );
-               FC_ASSERT( privkey );
-               keys[ addrs.back() ] = *privkey;
-            }
-            else
-            {
-               wlog( "Somehow _keys has no private key for extra_keys public key ${k}", ("k", pub) );
-            }
-         }
-         has_wildcard = true;
-      }
-      else
-      {
-         optional< private_key_type > key = wif_to_key( wif_key );
-         FC_ASSERT( key.valid(), "Invalid private key" );
-         fc::ecc::public_key pk = key->get_public_key();
-         addrs.push_back( pk );
-         keys[addrs.back()] = *key;
-         // see chain/balance_evaluator.cpp
-         addrs.push_back( pts_address( pk, false, 56 ) );
-         keys[addrs.back()] = *key;
-         addrs.push_back( pts_address( pk, true, 56 ) );
-         keys[addrs.back()] = *key;
-         addrs.push_back( pts_address( pk, false, 0 ) );
-         keys[addrs.back()] = *key;
-         addrs.push_back( pts_address( pk, true, 0 ) );
-         keys[addrs.back()] = *key;
-      }
-   }
-
-   vector< balance_object > balances = _remote_db->get_balance_objects( addrs );
-   addrs.clear();
-
-   set<asset_id_type> bal_types;
-   for( auto b : balances ) bal_types.insert( b.balance.asset_id );
-
-   struct claim_tx
-   {
-      vector< balance_claim_operation > ops;
-      set< address > addrs;
-   };
-   vector< claim_tx > claim_txs;
-
-   for( const asset_id_type& a : bal_types )
-   {
-      balance_claim_operation op;
-      op.deposit_to_account = claimer.id;
-      for( const balance_object& b : balances )
-      {
-         if( b.balance.asset_id == a )
-         {
-            op.total_claimed = b.available( dpo.time );
-            if( op.total_claimed.amount == 0 )
-               continue;
-            op.balance_to_claim = b.id;
-            op.balance_owner_key = keys[b.owner].get_public_key();
-            if( (claim_txs.empty()) || (claim_txs.back().ops.size() >= max_ops_per_tx) )
-               claim_txs.emplace_back();
-            claim_txs.back().ops.push_back(op);
-            claim_txs.back().addrs.insert(b.owner);
-         }
-      }
-   }
-
-   vector< signed_transaction > result;
-
-   for( const claim_tx& ctx : claim_txs )
-   {
-      signed_transaction tx;
-      tx.operations.reserve( ctx.ops.size() );
-      for( const balance_claim_operation& op : ctx.ops )
-         tx.operations.emplace_back( op );
-      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees() );
-      tx.validate();
-      signed_transaction signed_tx = sign_transaction( tx, false );
-      for( const address& addr : ctx.addrs )
-         signed_tx.sign( keys[addr], _chain_id );
-      // if the key for a balance object was the same as a key for the account we're importing it into,
-      // we may end up with duplicate signatures, so remove those
-      boost::erase(signed_tx.signatures, boost::unique<boost::return_found_end>(boost::sort(signed_tx.signatures)));
-      result.push_back( signed_tx );
-      if( broadcast )
-         _remote_net_broadcast->broadcast_transaction(signed_tx);
-   }
-
-   return result;
-} FC_CAPTURE_AND_RETHROW( (name_or_id) ) }
-
-}
-
 map<public_key_type, string> wallet_api::dump_private_keys()
 {
    FC_ASSERT(!is_locked());
