+++ libraries/wallet/wallet_asset.cpp
@@ -1,0 +1,354 @@
+/*
+ * Copyright (c) 2017 Cryptonomex, Inc., and contributors.
+ *
+ * The MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+#include <graphene/wallet/wallet_api_impl.hpp>
+#include <graphene/wallet/wallet.hpp>
+
+namespace graphene { namespace wallet { namespace detail {
+
+   std::string wallet_api_impl::asset_id_to_string(asset_id_type id) const
+   {
+      std::string asset_id = fc::to_string(id.space_id) +
+                             "." + fc::to_string(id.type_id) +
+                             "." + fc::to_string(id.instance.value);
+      return asset_id;
+   }
+
+   optional<extended_asset_object> wallet_api_impl::find_asset(asset_id_type id)const
+   {
+      auto rec = _remote_db->get_assets({asset_id_to_string(id)}, {}).front();
+      return rec;
+   }
+
+   optional<extended_asset_object> wallet_api_impl::find_asset(string asset_symbol_or_id)const
+   {
+      FC_ASSERT( asset_symbol_or_id.size() > 0 );
+
+      if( auto id = maybe_id<asset_id_type>(asset_symbol_or_id) )
+      {
+         // It's an ID
+         return find_asset(*id);
+      } else {
+         // It's a symbol
+         auto rec = _remote_db->lookup_asset_symbols({asset_symbol_or_id}).front();
+         if( rec )
+         {
+            if( rec->symbol != asset_symbol_or_id )
+               return optional<asset_object>();
+         }
+         return rec;
+      }
+   }
+
+   extended_asset_object wallet_api_impl::get_asset(asset_id_type id)const
+   {
+      auto opt = find_asset(id);
+      FC_ASSERT(opt);
+      return *opt;
+   }
+
+   extended_asset_object wallet_api_impl::get_asset(string asset_symbol_or_id)const
+   {
+      auto opt = find_asset(asset_symbol_or_id);
+      FC_ASSERT(opt);
+      return *opt;
+   }
+
+   asset_id_type wallet_api_impl::get_asset_id(string asset_symbol_or_id) const
+   {
+      FC_ASSERT( asset_symbol_or_id.size() > 0 );
+      vector<optional<extended_asset_object>> opt_asset;
+      if( std::isdigit( asset_symbol_or_id.front() ) )
+         return fc::variant(asset_symbol_or_id, 1).as<asset_id_type>( 1 );
+      opt_asset = _remote_db->lookup_asset_symbols( {asset_symbol_or_id} );
+      FC_ASSERT( (opt_asset.size() > 0) && (opt_asset[0].valid()) );
+      return opt_asset[0]->id;
+   }
+
+   signed_transaction wallet_api_impl::create_asset(string issuer, string symbol,
+         uint8_t precision, asset_options common, fc::optional<bitasset_options> bitasset_opts,
+         bool broadcast )
+   { try {
+      account_object issuer_account = get_account( issuer );
+      FC_ASSERT(!find_asset(symbol).valid(), "Asset with that symbol already exists!");
+
+      asset_create_operation create_op;
+      create_op.issuer = issuer_account.id;
+      create_op.symbol = symbol;
+      create_op.precision = precision;
+      create_op.common_options = common;
+      create_op.bitasset_opts = bitasset_opts;
+
+      signed_transaction tx;
+      tx.operations.push_back( create_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (issuer)(symbol)(precision)(common)(bitasset_opts)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::update_asset(string symbol, optional<string> new_issuer,
+         asset_options new_options, bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_update = find_asset(symbol);
+      if (!asset_to_update)
+         FC_THROW("No asset with that symbol exists!");
+      optional<account_id_type> new_issuer_account_id;
+      if (new_issuer)
+      {
+         FC_THROW( "The use of 'new_issuer' is no longer supported. Please use `update_asset_issuer' instead!" );
+      }
+
+      asset_update_operation update_op;
+      update_op.issuer = asset_to_update->issuer;
+      update_op.asset_to_update = asset_to_update->id;
+      update_op.new_issuer = new_issuer_account_id;
+      update_op.new_options = new_options;
+
+      signed_transaction tx;
+      tx.operations.push_back( update_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(new_issuer)(new_options)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::update_asset_issuer(string symbol, string new_issuer,
+         bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_update = find_asset(symbol);
+      if (!asset_to_update)
+        FC_THROW("No asset with that symbol exists!");
+
+      account_object new_issuer_account = get_account(new_issuer);
+
+      asset_update_issuer_operation update_issuer;
+      update_issuer.issuer = asset_to_update->issuer;
+      update_issuer.asset_to_update = asset_to_update->id;
+      update_issuer.new_issuer = new_issuer_account.id;
+
+      signed_transaction tx;
+      tx.operations.push_back( update_issuer );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(new_issuer)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::update_bitasset(string symbol, bitasset_options new_options,
+         bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_update = find_asset(symbol);
+      if (!asset_to_update)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_update_bitasset_operation update_op;
+      update_op.issuer = asset_to_update->issuer;
+      update_op.asset_to_update = asset_to_update->id;
+      update_op.new_options = new_options;
+
+      signed_transaction tx;
+      tx.operations.push_back( update_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(new_options)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::update_asset_feed_producers(string symbol, 
+         flat_set<string> new_feed_producers, bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_update = find_asset(symbol);
+      if (!asset_to_update)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_update_feed_producers_operation update_op;
+      update_op.issuer = asset_to_update->issuer;
+      update_op.asset_to_update = asset_to_update->id;
+      update_op.new_feed_producers.reserve(new_feed_producers.size());
+      std::transform(new_feed_producers.begin(), new_feed_producers.end(),
+                     std::inserter(update_op.new_feed_producers, update_op.new_feed_producers.end()),
+                     [this](const std::string& account_name_or_id){ return get_account_id(account_name_or_id); });
+
+      signed_transaction tx;
+      tx.operations.push_back( update_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(new_feed_producers)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::publish_asset_feed(string publishing_account, string symbol,
+         price_feed feed, bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_update = find_asset(symbol);
+      if (!asset_to_update)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_publish_feed_operation publish_op;
+      publish_op.publisher = get_account_id(publishing_account);
+      publish_op.asset_id = asset_to_update->id;
+      publish_op.feed = feed;
+
+      signed_transaction tx;
+      tx.operations.push_back( publish_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (publishing_account)(symbol)(feed)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::fund_asset_fee_pool(string from, string symbol, string amount,
+         bool broadcast /* = false */)
+   { try {
+      account_object from_account = get_account(from);
+      optional<asset_object> asset_to_fund = find_asset(symbol);
+      if (!asset_to_fund)
+        FC_THROW("No asset with that symbol exists!");
+      asset_object core_asset = get_asset(asset_id_type());
+
+      asset_fund_fee_pool_operation fund_op;
+      fund_op.from_account = from_account.id;
+      fund_op.asset_id = asset_to_fund->id;
+      fund_op.amount = core_asset.amount_from_string(amount).amount;
+
+      signed_transaction tx;
+      tx.operations.push_back( fund_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (from)(symbol)(amount)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::claim_asset_fee_pool(string symbol, string amount,
+         bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_pool_to_claim = find_asset(symbol);
+      if (!asset_pool_to_claim)
+        FC_THROW("No asset with that symbol exists!");
+      asset_object core_asset = get_asset(asset_id_type());
+
+      asset_claim_pool_operation claim_op;
+      claim_op.issuer = asset_pool_to_claim->issuer;
+      claim_op.asset_id = asset_pool_to_claim->id;
+      claim_op.amount_to_claim = core_asset.amount_from_string(amount).amount;
+
+      signed_transaction tx;
+      tx.operations.push_back( claim_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(amount)(broadcast) ) }
+
+
+   signed_transaction wallet_api_impl::reserve_asset(string from, string amount, string symbol,
+         bool broadcast /* = false */)
+   { try {
+      account_object from_account = get_account(from);
+      optional<asset_object> asset_to_reserve = find_asset(symbol);
+      if (!asset_to_reserve)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_reserve_operation reserve_op;
+      reserve_op.payer = from_account.id;
+      reserve_op.amount_to_reserve = asset_to_reserve->amount_from_string(amount);
+
+      signed_transaction tx;
+      tx.operations.push_back( reserve_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (from)(amount)(symbol)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::global_settle_asset(string symbol, price settle_price,
+         bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_settle = find_asset(symbol);
+      if (!asset_to_settle)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_global_settle_operation settle_op;
+      settle_op.issuer = asset_to_settle->issuer;
+      settle_op.asset_to_settle = asset_to_settle->id;
+      settle_op.settle_price = settle_price;
+
+      signed_transaction tx;
+      tx.operations.push_back( settle_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (symbol)(settle_price)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::settle_asset(string account_to_settle, string amount_to_settle,
+         string symbol, bool broadcast /* = false */)
+   { try {
+      optional<asset_object> asset_to_settle = find_asset(symbol);
+      if (!asset_to_settle)
+        FC_THROW("No asset with that symbol exists!");
+
+      asset_settle_operation settle_op;
+      settle_op.account = get_account_id(account_to_settle);
+      settle_op.amount = asset_to_settle->amount_from_string(amount_to_settle);
+
+      signed_transaction tx;
+      tx.operations.push_back( settle_op );
+      set_operation_fees( tx, _remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction( tx, broadcast );
+   } FC_CAPTURE_AND_RETHROW( (account_to_settle)(amount_to_settle)(symbol)(broadcast) ) }
+
+   signed_transaction wallet_api_impl::issue_asset(string to_account, string amount, string symbol,
+         string memo, bool broadcast )
+   {
+      auto asset_obj = get_asset(symbol);
+
+      account_object to = get_account(to_account);
+      account_object issuer = get_account(asset_obj.issuer);
+
+      asset_issue_operation issue_op;
+      issue_op.issuer           = asset_obj.issuer;
+      issue_op.asset_to_issue   = asset_obj.amount_from_string(amount);
+      issue_op.issue_to_account = to.id;
+
+      if( memo.size() )
+      {
+         issue_op.memo = memo_data();
+         issue_op.memo->from = issuer.options.memo_key;
+         issue_op.memo->to = to.options.memo_key;
+         issue_op.memo->set_message(get_private_key(issuer.options.memo_key),
+                                    to.options.memo_key, memo);
+      }
+
+      signed_transaction tx;
+      tx.operations.push_back(issue_op);
+      set_operation_fees(tx,_remote_db->get_global_properties().parameters.get_current_fees());
+      tx.validate();
+
+      return sign_transaction(tx, broadcast);
+   }
+
+}}} // namespace graphene::wallet::detail
